Created migration for IntialUser
I used the 'mix ecto.gen.migration IntialUser' to create a migration, I then added the keys we wish to store in the database. Then using mix ecto.migrate to apply these changes to the database. My nplan next is to create the schema for the user. This inital user migration can later be extended with further migrations