added User schema and starting changeset
Schemas are smiliar to Structs, the password field has a virtual key which means this key will never be stored in the database. You can use the Schema in this way:

alias Comp6000.Schemas.User
alias Comp6000.Repo

new_user = %User{username: Johnny1223}
Repo.insert(User.changeset(new_user, %{}))

Changesets are a way to validate, filter and apply constraints to the data provided. Currently this change set simply turns the 5 keys in the cast to a changeset and then ensures the username is unique. Later on this will most likely need to have various constrains such as username length etc.